% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Andwd}
\alias{Andwd}
\title{An Configuration Matrix for Distance Weighted Discrimination}
\format{A matrix with 50 rows and 3 columns}
\usage{
data(Andwd)
}
\description{
An Configuration Matrix for Distance Weighted Discrimination
}
\keyword{datasets}
