% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Apdwd}
\alias{Apdwd}
\title{Ap Configuration Matrix for Distance Weighted Discrimination}
\format{A matrix with 50 rows and 3 columns}
\usage{
data(Apdwd)
}
\description{
Ap Configuration Matrix for Distance Weighted Discrimination
}
\keyword{datasets}
