% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Betp}
\alias{Betp}
\title{Symmetric Matrix for Educational Testing Problem}
\format{A matrix with 5 rows and 5 columns}
\usage{
data(Betp)
}
\description{
Symmetric Matrix for Educational Testing Problem
}
\keyword{datasets}
