% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Bgpp}
\alias{Bgpp}
\title{Adjacency Matrix for Graph Partitioning Problem}
\format{A matrix with 10 rows and 10 columns}
\usage{
data(Bgpp)
}
\description{
Adjacency Matrix for Graph Partitioning Problem
}
\keyword{datasets}
