% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Blogcheby}
\alias{Blogcheby}
\title{B Matrix for the Log Chebyshev Approximation Problem}
\format{A matrix with 10 rows and 10 columns}
\usage{
data(Blogcheby)
}
\description{
B Matrix for the Log Chebyshev Approximation Problem
}
\keyword{datasets}
