% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Bmaxcut}
\alias{Bmaxcut}
\title{Adjacency Matrix for Max-Cut}
\format{A matrix with 10 rows and 10 columns}
\usage{
data(Bmaxcut)
}
\description{
Adjacency Matrix for Max-Cut
}
\keyword{datasets}
