% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Bmaxkcut}
\alias{Bmaxkcut}
\title{Adjacency Matrix for Max-kCut}
\format{A matrix with 10 rows and 10 columns}
\usage{
data(Bmaxkcut)
}
\description{
Adjacency Matrix for Max-kCut
}
\keyword{datasets}
