% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_theory.R
\name{control_theory}
\alias{control_theory}
\title{Control Theory}
\usage{
control_theory(B)
}
\arguments{
\item{B}{a matrix object containing square matrices of size n}
}
\value{
\item{X}{A list containing the solution matrix to the primal problem}
\item{y}{A list containing the  solution vector to the dual problem}
\item{Z}{A list containing the  solution matrix to the dual problem}
\item{pobj}{The achieved value of the primary objective function}
\item{dobj}{The achieved value of the dual objective function}
}
\description{
\code{control_theory} creates input for sqlp to solve the Control Theory Problem
}
\details{
Solves the control theory problem. Mathematical and implementation
details can be found in the vignette
}
\examples{
B <- matrix(list(),2,1)
B[[1]] <- matrix(c(-.8,1.2,-.5,-1.1,-1,-2.5,2,.2,-1),nrow=3,byrow=TRUE)
B[[2]] <- matrix(c(-1.5,.5,-2,1.1,-2,.2,-1.4,1.1,-1.5),nrow=3,byrow=TRUE)

out <- control_theory(B)

}
