% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{DoptDesign}
\alias{DoptDesign}
\title{Test Vector Matrix for D-Optimal Design}
\format{A matrix with 3 rows and 25 columns}
\usage{
data(DoptDesign)
}
\description{
Test Vector Matrix for D-Optimal Design
}
\keyword{datasets}
