% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{flogcheby}
\alias{flogcheby}
\title{f vector for the Log Chebyshev Approximation Problem}
\format{A vector with length 20}
\usage{
data(flogcheby)
}
\description{
f vector for the Log Chebyshev Approximation Problem
}
\keyword{datasets}
