% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Ftoep}
\alias{Ftoep}
\title{Symmetric Matrix for the Toeplitz Approximatin Problem}
\format{A matrix with 10 rows and 10 columns}
\usage{
data(Ftoep)
}
\description{
Symmetric Matrix for the Toeplitz Approximatin Problem
}
\keyword{datasets}
