% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Glovasz}
\alias{Glovasz}
\title{Adjacency Matrix on which to find the Lovasz Number}
\format{A matrix with 10 rows and 10 columns}
\usage{
data(Glovasz)
}
\description{
Adjacency Matrix on which to find the Lovasz Number
}
\keyword{datasets}
