% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Hnearcorr}
\alias{Hnearcorr}
\title{Approximate Correlation Matrix for Nearest Correlation Matrix Problem}
\format{A matrix with 5 rows and 5 columns}
\usage{
data(Hnearcorr)
}
\description{
Approximate Correlation Matrix for Nearest Correlation Matrix Problem
}
\keyword{datasets}
