% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmi.R
\name{lmi1}
\alias{lmi1}
\title{Linear Matrix Inequality 1}
\usage{
lmi1(B)
}
\arguments{
\item{B}{An mxn real valued matrix}
}
\value{
\item{X}{A list containing the solution matrix to the primal problem}
\item{y}{A list containing the  solution vector to the dual problem}
\item{Z}{A list containing the  solution matrix to the dual problem}
\item{pobj}{The achieved value of the primary objective function}
\item{dobj}{The achieved value of the dual objective function}
}
\description{
\code{lmi1} creates input for sqlp to solve a linear matrix inequality problem
}
\details{
Solves the type-1 linear matrix inequality problem. Mathematical and implementation
details can be found in the vignette
}
\examples{
B <- matrix(c(-1,5,1,0,-2,1,0,0,-1), nrow=3)

#Not Run
#out <- lmi1(B)

}
