% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logcheby.R
\name{logcheby}
\alias{logcheby}
\title{Log Chebyshev Approximation}
\usage{
logcheby(B, f)
}
\arguments{
\item{B}{A pxm real valued matrix with p > m}

\item{f}{A vector of length p}
}
\value{
\item{X}{A list containing the solution matrix to the primal problem}
\item{y}{A list containing the  solution vector to the dual problem}
\item{Z}{A list containing the  solution matrix to the dual problem}
\item{pobj}{The achieved value of the primary objective function}
\item{dobj}{The achieved value of the dual objective function}
}
\description{
\code{logcheby} creates input for sqlp to solve the Chebyshev Approximation Problem
}
\details{
Solves the log Chebyshev approximation problem. Mathematical and implementation
details can be found in the vignette
}
\examples{
data(Blogcheby)
data(flogcheby)

#Not Run
#out <- logcheby(Blogcheby, flogcheby)

}
