% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovasz.R
\name{lovasz}
\alias{lovasz}
\title{Lovasz Number of a Graph}
\usage{
lovasz(G)
}
\arguments{
\item{G}{An adjacency matrix corresponding to a graph}
}
\value{
\item{X}{A list containing the solution matrix to the primal problem}
\item{y}{A list containing the  solution vector to the dual problem}
\item{Z}{A list containing the  solution matrix to the dual problem}
\item{pobj}{The achieved value of the primary objective function}
\item{dobj}{The achieved value of the dual objective function}
}
\description{
\code{lovasz} creates input for sqlp to find the Lovasz Number of a graph
}
\details{
Finds the maximum Shannon entropy of a graph, more commonly known as the Lovasz number. 
Mathematical and implementation details can be found in the vignette
}
\examples{
data(Glovasz)

out <- lovasz(Glovasz)

}
