% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Vminelips}
\alias{Vminelips}
\title{Configuration Matrix for Minimum Ellipse Problem}
\format{A matrix with 2 rows and 2 columns}
\usage{
data(Vminelips)
}
\description{
Configuration Matrix for Minimum Ellipse Problem
}
\keyword{datasets}
